refactos shuffle_writer
